requirements-minimal-working.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
Pillow==10.1.0
python-json-logger==2.0.7
python-dotenv==1.0.0
aiofiles==23.2.1



